+++ datasets/tr/wanted/tr_wanted.yml
@@ -6,6 +6,7 @@ coverage:
   frequency: daily
   start: 2024-04-02
 load_db_uri: ${OPENSANCTIONS_DATABASE_URI}
+url: https://en.terorarananlar.pol.tr/
 summary: >
   List of people considered wanted for terrorism by the Türkiye Ministry of Interior
 description: |
